Refactor and simplify wav example and rename to play_wav for clarity
